include iOS < 10 notification handling even when compiling with 10+ SDK
Current Xcode ships with iOS 10 SDK. This commit fixes an issue where if
you compile a project with 10 SDK and run on iOS 9, the incoming
notifications would not be handled. This is because the compile-time
conditionals would exclude methods needed for handling incoming
notifications on iOS 9. So always include the methods for iOS < 10, and
just don't execute them if we compiled and actually run on iOS 10+.